requirements.txt
asgiref==3.2.10
dataclasses==0.7
Django==3.0.5
django-mptt==0.11.0
gunicorn==20.0.4
numpy==1.19.2
pytz==2020.1
sqlparse==0.2.4
whitenoise==5.2.0
